feat: split functions into separate files
