+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "prettier.configPath": "C:\\Users\\Lenovo\\Desktop\\SPAralige\\.prettierrc.json"
+  "prettier.configPath": "C:\\Users\\Nitro5-R\\AppData\\Roaming\\Code\\User\\settings.json"
 }
